+++ src/bmu/models/__init__.py
@@ -1,4 +1,5 @@
 from bmu.models.credential import Credential, CredentialProvider
+from bmu.models.cve_scan import CveResult, CveScan
 from bmu.models.device import Device
 from bmu.models.group import DeviceGroup
 from bmu.models.parser_template import ParserTemplate, ParserType
@@ -9,6 +10,8 @@ from bmu.models.schedule import JobKind, Schedule
 __all__ = [
     "Credential",
     "CredentialProvider",
+    "CveResult",
+    "CveScan",
     "Device",
     "DeviceGroup",
     "JobKind",